fix delete auth_key when logout application

+++ app/src/main/java/com/laoschool/screen/HomeActivity.java
@@ -694,21 +694,22 @@ public class HomeActivity extends AppCompatActivity implements
 
     @Override
     public void logoutApplication() {
-        SharedPreferences mySPrefs = PreferenceManager.getDefaultSharedPreferences(this);
+        SharedPreferences mySPrefs = this.getSharedPreferences(
+                LaoSchoolShared.SHARED_PREFERENCES_TAG, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = mySPrefs.edit();
         editor.remove("auth_key");
-        editor.apply();
 
-        //clear cache data
-        DataAccessMessage.deleteTable();
-        DataAccessImage.deleteTable();
+        if (editor.commit()) {
+            //clear cache data
+            DataAccessMessage.deleteTable();
+            DataAccessImage.deleteTable();
 
-        //
-        Intent intent = new Intent(this, ScreenLogin.class);
-        startActivity(intent);
-        finish();
-
-
+            Intent intent = new Intent(this, ScreenLogin.class);
+            startActivity(intent);
+            finish();
+        } else {
+            Toast.makeText(this, "Delete auth_key fails", Toast.LENGTH_LONG).show();
+        }
     }
 
     @Override
